Move WorkManager initialization to Application#onCreate
