+++ Covid19Radar/Covid19Radar.Android/MainActivity.cs
@@ -8,9 +8,6 @@ using Android.OS;
 using Android.Runtime;
 using Android.Content;
 using Acr.UserDialogs;
-using Xamarin.ExposureNotifications;
-using System;
-using AndroidX.Work;
 
 namespace Covid19Radar.Droid
 {
@@ -25,16 +22,6 @@ namespace Covid19Radar.Droid
             ToolbarResource = Resource.Layout.Toolbar;
             base.SetTheme(Resource.Style.MainTheme);
             base.OnCreate(savedInstanceState);
-
-            // Override WorkRequest configuration
-            // Must be run before being scheduled with `ExposureNotification.Init()` in `App.OnInitialized()`
-            var repeatInterval = TimeSpan.FromHours(6);
-            Action<PeriodicWorkRequest.Builder> requestBuilder = b =>
-               b.SetConstraints(new Constraints.Builder()
-                   .SetRequiresBatteryNotLow(true)
-                   .SetRequiredNetworkType(NetworkType.Connected)
-                   .Build());
-            ExposureNotification.ConfigureBackgroundWorkRequest(repeatInterval, requestBuilder);
 
             Xamarin.Forms.Forms.SetFlags("RadioButton_Experimental");
             Xamarin.Essentials.Platform.Init(this, savedInstanceState);
