+++ Covid19Radar/Covid19Radar.iOS/AppDelegate.cs
@@ -36,8 +36,8 @@ namespace Covid19Radar.iOS
         {
             NSUrlCache.SharedCache.RemoveAllCachedResponses();
 
+            App.InitExposureNotification();
             App.InitializeServiceLocator(RegisterPlatformTypes);
-            App.UseMockExposureNotificationImplementationIfNeeded();
 
             Xamarin.Forms.Forms.SetFlags("RadioButton_Experimental");
