+++ Covid19Radar/Covid19Radar/App.xaml.cs
@@ -17,6 +17,7 @@ using Covid19Radar.Services.Logs;
 using System;
 using CommonServiceLocator;
 using Covid19Radar.Common;
+using Xamarin.ExposureNotifications;
 
 /*
  * Our mission...is
@@ -50,8 +51,6 @@ namespace Covid19Radar
             LogFileService = Container.Resolve<ILogFileService>();
             LogFileService.AddSkipBackupAttribute();
 
-            Xamarin.ExposureNotifications.ExposureNotification.Init();
-
             // Local Notification tap event listener
             //NotificationCenter.Current.NotificationTapped += OnNotificationTapped;
             LogUnobservedTaskExceptions();
@@ -75,7 +74,14 @@ namespace Covid19Radar
             LoggerService.EndMethod();
         }
 
-        public static void UseMockExposureNotificationImplementationIfNeeded()
+        public static void InitExposureNotification()
+        {
+            UseMockExposureNotificationImplementationIfNeeded();
+
+            ExposureNotification.Init();
+        }
+
+        private static void UseMockExposureNotificationImplementationIfNeeded()
         {
 #if USE_MOCK
             // For debug mode, set the mock api provider to interact
